Fix -Pdist AOT training crash from the macOS Open-With handler (#290)
The dist AOT-training step (scripts/aot_build.java) launches the image's
own bin/java with a hand-built option list that did NOT include the
--add-exports the packaged launcher gets, so training crashed with
IllegalAccessError loading the Glass EventHandler subclass -> no AOT
cache shipped.

Two fixes:
- Move the handler out of App into a standalone top-level class
  com.editora.MacOpenFiles. The FX launcher reflects over App and
  eager-loads its nest members, so a com.sun.glass.ui subclass nested in
  App was class-loaded (and access-checked) before start() ran, outside
  the try/catch. As a separate class it loads only inside the guarded
  install(), degrading to a logged no-op if the export is ever missing.
- Add --add-exports=javafx.graphics/com.sun.glass.ui=com.editora to the
  aot_build.java training command so it mirrors the launcher.

Device-tested: -Pdist APP_IMAGE now trains successfully ("cache present:
editora.aot (62 MB)") with no IllegalAccessError; launcher maps the
cache cleanly; Info.plist still declares CFBundleDocumentTypes.
